Points are to be output in the order of their position on the board (left to right, top to bottom).
This allows us to get rid of the sort, but there are probably some cases
where our output is wrong: we visit all cells, keeping only the ones
with a single liberty. The coordinates of the lone liberty are output
in the order of the first cell of each group, instead of the actual
order of the liberty coordinates.